The knockback cell comes off the attack that threw it, and a struck body clears its own count by swinging
LIFE5's tier zero: the shared structures the rung's remaining phases stand on,
so each of them is a caller rather than a new mechanism.

The authored per-attack table replaces the stand-in. A contact now carries the
index of the swing record that landed it, and the cell is drawn from that
record's own four direction buckets — rotated by `+0xB8`, so bucket 0 is not
assumed to be any particular direction. The candidate NAME is the answer, which
is how the `SMALL` family and both `LOW_BACK` cells become reachable at all;
`StandInKnockbackSize`/`StandInKnockbackHeight` are gone, and so is
`FElysiumKnockback::Activity()`.

The eligibility gate is whole. The hit-buildup counter is one scalar on the
victim with retail's own four sites — raised by any attacker's landed hit,
cleared by the body's own swing reaching `melee_swing_completion_percent`, and
keyed to nobody — beside the `+0xBA == 2` unconditional marker and the
`npc_VTzimisceRunner` class bypass that skips the two terms under it.

Two orderings are recovered rather than chosen, both read off
`CBaseCombatWeapon::FUN_102579F0`. The increment runs before the test, so the
counter a blow tests already carries that blow and the default of 2 admits two
knockbacks. And the knockback runs AFTER the health commit where the blocked
reactions run before it — which is the whole of why a killing blow is never
thrown. The selector's own shape is recovered too: the draw is unconditional
even over one candidate, an empty bucket re-reads bucket 0, and the no-record
`0x8B` cell is a different fallback from the per-class default this runtime
does not carry. All of it is recorded in the owning doc.

The pipeline carries the two melee fields the cast arm will need: the reach
band's near edge at `+0x2CC`, whose unset marker is `FLT_MIN` and not `reach`'s
`FLT_MAX`, and the 24-byte attack envelopes at `+0x2BC`/`+0x2C0`. The envelope
corners take a scale and nothing else — their axes are reach distance, lateral
tolerance and vertical offset, so the positional path would mirror an axis that
is symmetric about zero and disagree with nothing. They ride sidecar columns 11
and 12 under the existing trailing-and-optional rule, so an un-re-exported
corpus reads them as authored absences and needs no manifest bump.

`IElysiumNpcMotor` gains the ballistic pair the flying chain will need:
`Launch` as an execution and `SampleBallistic` as a polled query, both
defaulted, both stating their headless answer. Retail assigns a velocity to a
live animated body rather than solving physics, so this is not the ragdoll
impulse vocabulary.

Alongside: the nine flying activity codes and their state projection, which
uncovered a `uint32` code mask that the vocabulary had already outgrown and a
whole-enum test walk that had silently stopped at `DieRagdoll`; a Substrate
melee-contact fixture that binds a fabricated `Melee_Reactions` block, so
`.Producer` drives real contacts instead of re-composing them; a swing-contact
test double that actually sweeps geometry; the 32x32x72 stand hull collapsed
onto the constants three queries were re-declaring; and a `save_png` that
survives two workers racing one shared basetexture, which is a POSIX guarantee
this pipeline was relying on from Windows.

Substrate 377/377, pipeline suite 73/73. The corpus re-exported and the two new
columns land on 46,726 clip rows.

+++ Source/ElysiumUE/Private/Debug/ElysiumLocomotionTrace.h
@@ -47,7 +47,12 @@ namespace ElysiumLocomotionTrace
 		int32 ResolvedFrames = 0;
 		int32 FallbackFrames = 0;
 		// Which activity codes the course reached, as a bitmask over `EElysiumAnimActivityCode`.
-		uint32 CodesSeen = 0;
+		//
+		// **64 bits because the enum is wider than 32.** The vocabulary carries the locomotion slice,
+		// the death pair, the ten grounded knockback cells and the nine flying ones, which is past
+		// ordinal 31 — and `1u << 32` is undefined behaviour, not a wrapped zero. A code added to the
+		// enum has to fit here, so the two are checked against each other at compile time below.
+		uint64 CodesSeen = 0;
 		// Fastest horizontal speed the course reached, Source units/s.
 		double PeakSpeed2D = 0.0;
 		// The body the selections were resolved for, and the banks they came out of.
